removed unnecessary (int)

+++ TheKoans/AboutArrays.cs
@@ -30,7 +30,7 @@ namespace TheKoans
             CollectionAssert.AreEqual(new int[] { 42 }, array, "These arrays are literally equal... But you won't see this string in the error message.");
 
             //Are arrays 0-based or 1-based?
-            Assert.AreEqual(42, array[((int)0)], "Well, it's either 0 or 1.. you have a 110010-110010 chance of getting it right.");
+            Assert.AreEqual(42, array[0], "Well, it's either 0 or 1.. you have a 110010-110010 chance of getting it right.");
 
             //This is important because...
             Assert.IsTrue(array.IsFixedSize, "...because Fixed Size arrays are not dynamic");
